review comments

+++ lib/include/public/LogSessionData.hpp
@@ -5,6 +5,7 @@
 #include "Version.hpp"
 
 #include <string>
+#include <cstdint>
 
 namespace ARIASDK_NS_BEGIN
 {
@@ -12,13 +13,13 @@ namespace ARIASDK_NS_BEGIN
     class LogSessionData
     {
     public:
-        LogSessionData(unsigned long long, std::string);
+        LogSessionData(uint64_t, std::string);
 
         /// <summary>
         /// Gets the time that this session began.
         /// </summary>
         /// <returns>A 64-bit integer that contains the time.</returns>
-        unsigned long long getSessionFirstTime() const;
+        uint64_t getSessionFirstTime() const;
 
         /// <summary>
         /// Gets the SDK unique identifier.
@@ -26,8 +27,8 @@ namespace ARIASDK_NS_BEGIN
         std::string getSessionSDKUid() const;
 
     protected:
-        unsigned long long                  m_sessionFirstTimeLaunch{0ull};
-        std::string                         m_sessionSDKUid;
+        uint64_t        m_sessionFirstTimeLaunch{0ull};
+        std::string     m_sessionSDKUid;
     };
 
 } ARIASDK_NS_END
